Added 'HelpInfo' object and moved 'shared_featurs.py' to tools. Added 'help_info' classmethod for BaseCommand.
