+++ source/tools/shared_features.py
@@ -7,6 +7,22 @@ from discord.utils import MISSING
 from source.tools.ui_helper import generate_embed
 
 
+class HelpInfo:
+    '''
+    This is simply a container for all the info that the help command needs.
+
+    # Attributes
+      `name`: Name of the command.
+      `desc`: Description of the command.
+      `group`: The group that the command belongs to (for slash commands). Defaults to None.
+      `mod_only`: Whether the command is mod_only. Defaults to False.
+    '''
+    def __init__(self, *, name: str, desc: str, group: str = None, mod_only: bool = False) -> None:
+        self.name = name
+        self.desc = desc
+        self.group = group
+        self.mod_only = mod_only
+
 class SupportModal(ui.Modal, title='Help Form'):
     '''
     Support Modal used by the help button and help command.
